+++ docs/examples/iconbuttonlink/enabled.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Flex, IconButtonLink } from 'gestalt';
 
 export default function Example() {
@@ -5,14 +6,13 @@ export default function Example() {
     <Flex alignItems="center" height="100%" justifyContent="center" width="100%">
       <IconButtonLink
         accessibilityLabel="Visit the Gestalt documentation"
-        href="https://gestalt.pinterest.systems"
+        href="https://gestalt.pinterest.systems/web/iconbuttonlink#Active-item"
         icon="visit"
         onClick={({ event, dangerouslyDisableOnNavigation }) => {
           event.preventDefault();
           dangerouslyDisableOnNavigation();
         }}
         target="blank"
-        tooltip={{ text: 'Besuchen Sie Pinterest' }}
       />
     </Flex>
   );